backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
playwright==1.44.0
tf-playwright-stealth
openai==1.30.0
apify-client==1.7.0
redis==5.0.4
pydantic==2.7.0
pydantic-settings==2.2.1
python-multipart==0.0.9
pytest==8.2.0
pytest-asyncio==0.23.6
